+++ assets/ButtonPage.27651b46.js
@@ -1,0 +1,13 @@
+var i=Object.defineProperty;var o=(e,r)=>i(e,"name",{value:r,configurable:!0});import{d as t,h as n}from"./index.4c2eaa26.js";import{C as u}from"./ComponentInfo.499e78fa.js";import{S as c}from"./copyText.e62e7e6d.js";import"./Link.32e525d3.js";import"./PaperCode.1e0fbfc5.js";function a(){return t(c,{spacing:2,direction:"row",get children(){return[t(n,{variant:"text",children:"Text"}),t(n,{variant:"contained",children:"Contained"}),t(n,{variant:"outlined",children:"Outlined"})]}})}o(a,"BasicButtons");a.code=`import Button from "@suid/material/Button";
+import Stack from "@suid/material/Stack";
+
+export default function BasicButtons() {
+  return (
+    <Stack spacing={2} direction="row">
+      <Button variant="text">Text</Button>
+      <Button variant="contained">Contained</Button>
+      <Button variant="outlined">Outlined</Button>
+    </Stack>
+  );
+}
+`;function f(){return t(u,{get name(){return n.name},docsName:"buttons",examples:[a]})}o(f,"ButtonPage");export{f as default};