+++ assets/CircularProgressPage.f94b1da5.js
@@ -1,0 +1,13 @@
+var t=Object.defineProperty;var e=(n,s)=>t(n,"name",{value:s,configurable:!0});import{C as o}from"./CircularProgress.443da225.js";import{C as a}from"./ComponentInfo.499e78fa.js";import{S as i}from"./copyText.e62e7e6d.js";import{d as r}from"./index.4c2eaa26.js";import"./Link.32e525d3.js";import"./PaperCode.1e0fbfc5.js";function c(){return r(i,{sx:{color:"grey.500"},spacing:2,direction:"row",get children(){return[r(o,{color:"secondary"}),r(o,{color:"success"}),r(o,{color:"inherit"})]}})}e(c,"CircularColor");c.code=`import CircularProgress from "@suid/material/CircularProgress";
+import Stack from "@suid/material/Stack";
+
+export default function CircularColor() {
+  return (
+    <Stack sx={{ color: "grey.500" }} spacing={2} direction="row">
+      <CircularProgress color="secondary" />
+      <CircularProgress color="success" />
+      <CircularProgress color="inherit" />
+    </Stack>
+  );
+}
+`;function f(){return r(a,{get name(){return o.name},docsName:"progress",examples:[c]})}e(f,"CircularProgressPage");export{f as default};