+++ assets/TextFieldPage.3ce6c651.js
@@ -1,0 +1,232 @@
+var Te=Object.defineProperty;var s=(e,n)=>Te(e,"name",{value:n,configurable:!0});import{q as R,r as $,s as F,v as P,w as h,K as b,o as _,f as K,a as ue,m as p,af as Re,n as y,d as o,A as W,az as G,a8 as le,aA as $e,am as ke,aB as ce,ae as Le,t as q,ai as J,S as pe,ab as ze,a6 as m,aC as Be,B as Q}from"./index.4c2eaa26.js";import{u as O,F as We}from"./useFormControl.29642d82.js";import{f as w,F as Pe}from"./formControlState.8cbfc4fb.js";import{u as Oe}from"./useControlled.d7e7dd38.js";import{C as we}from"./ComponentInfo.499e78fa.js";import"./Link.32e525d3.js";import"./copyText.e62e7e6d.js";import"./PaperCode.1e0fbfc5.js";function Ae(e,n,l,t,r){let a,i;return t!==r?(a=e.slice(0,t)+(l||"")+e.slice(r),i=t,n.startsWith("insert")&&i++):n.startsWith("insert")?(a=e.slice(0,t)+l+e.slice(r),i=t+l.length):n==="deleteContentBackward"?(i=Math.max(0,t-1),a=e.slice(0,i)+e.slice(t)):n==="deleteContentForward"?(a=e.slice(0,t)+e.slice(t+1),i=t):(a=e,i=t),[a,i]}s(Ae,"applyInputTransform");function Se(e){e.element.addEventListener("beforeinput",n=>{const l=n.currentTarget;if(n.inputType==="insertLineBreak")return;if(typeof l.selectionStart!="number")throw new Error("'selectionStart' can not be null");if(typeof l.selectionEnd!="number")throw new Error("'selectionEnd' can not be null");const[t,r]=Ae(l.value,n.inputType,n.data||"",l.selectionStart,l.selectionEnd);e.onChange(n,t,r),n.preventDefault()})}s(Se,"prepareControlledInput");function Me(e){const[n,l]=e.valueAndStart,t=l??n.length;e.element.value=n,e.element.setSelectionRange(t,t)}s(Me,"setControlledInputValue");function qe(e){return $("MuiInputBase",e)}s(qe,"getInputBaseUtilityClass");const j=R("MuiInputBase",["root","formControl","focused","disabled","adornedStart","adornedEnd","error","sizeSmall","multiline","colorSecondary","fullWidth","hiddenLabel","input","inputSizeSmall","inputMultiline","inputTypeSearch","inputAdornedStart","inputAdornedEnd","inputHiddenLabel"]);function re(e){return e!=null&&!(Array.isArray(e)&&e.length===0)}s(re,"hasValue");function Ne(e,n=!1){return e&&(re(e.value)&&e.value!==""||n&&re(e.defaultValue)&&e.defaultValue!=="")}s(Ne,"isFilled");const oe=F()({name:"MuiInputBase",propDefaults:({set:e})=>e({components:{},componentsProps:{},fullWidth:!1,inputComponent:"input",inputProps:{},multiline:!1,type:"text",disableInjectingGlobalStyles:!1}),selfPropNames:["aria-describedby","autoComplete","autoFocus","classes","color","components","componentsProps","defaultValue","disableInjectingGlobalStyles","disabled","endAdornment","error","fullWidth","id","inputComponent","inputProps","inputProps","inputRef","margin","maxRows","minRows","multiline","name","onBlur","onChange","onFocus","onKeyDown","onKeyUp","placeholder","readOnly","renderSuffix","required","rows","size","startAdornment","type","value"],utilityClass:qe,autoCallUseClasses:!1,slotClasses:e=>({root:["root",`color${P(e.color)}`,!!e.disabled&&"disabled",!!e.error&&"error",!!e.fullWidth&&"fullWidth",e.focused&&"focused",!!e.formControl&&"formControl",e.size==="small"&&"sizeSmall",e.multiline&&"multiline",!!e.startAdornment&&"adornedStart",!!e.endAdornment&&"adornedEnd",!!e.hiddenLabel&&"hiddenLabel"],input:["input",!!e.disabled&&"disabled",e.type==="search"&&"inputTypeSearch",e.multiline&&"inputMultiline",e.size==="small"&&"inputSizeSmall",!!e.hiddenLabel&&"inputHiddenLabel",!!e.startAdornment&&"inputAdornedStart",!!e.endAdornment&&"inputAdornedEnd"]})}),N=s((e,n)=>{const l=e.ownerState;return[n.root,!!l.formControl&&n.formControl,!!l.startAdornment&&n.adornedStart,!!l.endAdornment&&n.adornedEnd,!!l.error&&n.error,l.size==="small"&&n.sizeSmall,l.multiline&&n.multiline,l.color&&n[`color${P(l.color)}`],!!l.fullWidth&&n.fullWidth,!!l.hiddenLabel&&n.hiddenLabel]},"rootOverridesResolver"),V=s((e,n)=>{const l=e.ownerState;return[n.input,l.size==="small"&&n.inputSizeSmall,l.multiline&&n.inputMultiline,l.type==="search"&&n.inputTypeSearch,!!l.startAdornment&&n.inputAdornedStart,!!l.endAdornment&&n.inputAdornedEnd,!!l.hiddenLabel&&n.inputHiddenLabel]},"inputOverridesResolver"),H=h("div",{name:"MuiInputBase",slot:"Root",overridesResolver:N})(({theme:e,ownerState:n})=>({...e.typography.body1,color:e.palette.text.primary,lineHeight:"1.4375em",boxSizing:"border-box",position:"relative",cursor:"text",display:"inline-flex",alignItems:"center",[`&.${j.disabled}`]:{color:e.palette.text.disabled,cursor:"default"},...n.multiline&&{padding:"4px 0 5px",...n.size==="small"&&{paddingTop:1}},...n.fullWidth&&{width:"100%"}})),E=h("input",{name:"MuiInputBase",slot:"Input",overridesResolver:V})(({theme:e,ownerState:n})=>{const l=e.palette.mode==="light",t={color:"currentColor",opacity:l?.42:.5,transition:e.transitions.create("opacity",{duration:e.transitions.duration.shorter})},r={opacity:"0 !important"},a={opacity:l?.42:.5};return{font:"inherit",letterSpacing:"inherit",color:"currentColor",padding:"4px 0 5px",border:0,boxSizing:"content-box",background:"none",height:"1.4375em",margin:0,WebkitTapHighlightColor:"transparent",display:"block",minWidth:0,width:"100%",animationName:"mui-auto-fill-cancel",animationDuration:"10ms","&::-webkit-input-placeholder":t,"&::-moz-placeholder":t,"&:-ms-input-placeholder":t,"&::-ms-input-placeholder":t,"&:focus":{outline:0},"&:invalid":{boxShadow:"none"},"&::-webkit-search-decoration":{WebkitAppearance:"none"},[`label[data-shrink=false] + .${j.formControl} &`]:{"&::-webkit-input-placeholder":r,"&::-moz-placeholder":r,"&:-ms-input-placeholder":r,"&::-ms-input-placeholder":r,"&:focus::-webkit-input-placeholder":a,"&:focus::-moz-placeholder":a,"&:focus:-ms-input-placeholder":a,"&:focus::-ms-input-placeholder":a},[`&.${j.disabled}`]:{opacity:1,WebkitTextFillColor:e.palette.text.disabled},"&:-webkit-autofill":{animationDuration:"5000s",animationName:"mui-auto-fill"},...n.size==="small"&&{paddingTop:1},...n.multiline&&{height:"auto",resize:"none",padding:0,paddingTop:0},...n.type==="search"&&{MozAppearance:"textfield"}}}),Ve=s(()=>o($e,{styles:{"@keyframes mui-auto-fill":{from:{display:"block"}},"@keyframes mui-auto-fill-cancel":{from:{display:"block"}}}}),"inputGlobalStyles"),Y=oe.component(s(function({allProps:n,otherProps:l,props:t}){const r=b(()=>t.inputProps.value!=null?t.inputProps.value:t.value),a=(r()??null)!==null,[i,f]=Oe({controlled:()=>r(),default:()=>t.defaultValue,name:"InputBase"}),c=ke({ref:d=>{typeof t.inputRef=="function"&&t.inputRef(d)}});let k;a?_(()=>{Se({element:c.ref,onChange:(d,C,Ie)=>{k=Ie,typeof t.inputProps.onChange=="function"&&t.inputProps.onChange(d),f(C),typeof t.onChange=="function"&&t.onChange(d,C)}})}):_(()=>{c.ref.addEventListener("input",d=>{const C=c.ref.value;typeof t.inputProps.onChange=="function"&&t.inputProps.onChange(d),f(C),typeof t.onChange=="function"&&t.onChange(d,C)})}),K(d=>{if(a||d){const C=i();typeof C=="string"&&Me({element:c.ref,valueAndStart:[C,k]})}return!1},!0);const[I,z]=ue(!1),g=O(),x=w({props:n,muiFormControl:g,states:["color","disabled","error","hiddenLabel","size","required","filled"]}),v=p(()=>x,()=>({focused:g?g.focused:I()}));K(()=>{!g&&t.disabled&&I()&&(z(!1),typeof t.onBlur=="function"&&t.onBlur?.(null))});const ge=b(()=>g&&g.onFilled),he=b(()=>g&&g.onEmpty),A=s(d=>{Ne(d)?ge()?.():he()?.()},"checkDirty");Re(()=>{a&&A({value:i()})}),_(()=>{A(c.ref)});const Z=b(()=>t.multiline&&t.inputComponent==="input"),xe=b(()=>{const d=t.inputComponent;return Z(),d}),U=b(()=>{let d=t.inputProps;return Z()&&(t.rows?d={type:void 0,minRows:t.rows,maxRows:t.rows,...d}:d={type:void 0,maxRows:t.maxRows,minRows:t.minRows,...d}),p(d,()=>t.componentsProps.input||{})});K(()=>{g?.setAdornedStart(Boolean(t.startAdornment))});const S=p(n,()=>({...t,color:v.color||"primary",disabled:v.disabled,error:v.error,focused:v.focused,formControl:g,hiddenLabel:v.hiddenLabel,size:v.size})),ee=oe.useClasses(S),ne=b(()=>t.components.Root||H),D=b(()=>t.componentsProps.root||{}),te=b(()=>t.components.Input||E),ve=p(()=>S,()=>D().ownerState||{}),Ce=p(()=>S,()=>U().ownerState||{}),ye=p(v,()=>({startAdornment:t.startAdornment})),Fe=b(()=>t.renderSuffix?.(ye));return[y((()=>{const d=y(()=>!t.disableInjectingGlobalStyles,!0);return()=>d()&&Ve()})()),o(G,p(D,l,{get component(){return ne()}},()=>!le(ne())&&{ownerState:ve},{onClick:d=>{c.ref&&d.currentTarget===d.target&&c.ref.focus(),typeof l.onClick=="function"&&l.onClick(d)},get className(){return W(ee.root,D().className,l.className)},get children(){return[y(()=>t.startAdornment),o(We.Provider,{value:void 0,get children(){return o(G,p({get component(){return te()},ownerState:S,get["aria-invalid"](){return v.error},get["aria-describedby"](){return t["aria-describedby"]},get autocomplete(){return t.autoComplete},get autofocus(){return t.autoFocus},get disabled(){return v.disabled},get id(){return t.id},onAnimationStart:d=>{A(d.animationName==="mui-auto-fill-cancel"?c.ref:{value:"x"})},get name(){return t.name},get placeholder(){return t.placeholder},get readOnly(){return t.readOnly},get required(){return v.required},get rows(){return t.rows},get onKeyDown(){return t.onKeyDown},get onKeyUp(){return t.onKeyUp},get type(){return t.type}},U,()=>!le(te())&&{as:xe(),ownerState:Ce},{ref:c,get className(){return W(ee.input,U().className)},onBlur:d=>{t.onBlur?.(d),typeof t.inputProps.onBlur=="function"&&t.inputProps.onBlur(d),g&&g.onBlur?g.onBlur():z(!1)},onInput:d=>{if(!a){const C=d.target||c.ref;if(C==null)throw new Error("MUI: Expected valid input target. Did you use a custom `inputComponent` and forget to forward refs? See https://mui.com/r/input-component-ref-interface for more info.");A({value:C.value})}},onFocus:d=>{if(v.disabled){d.stopPropagation();return}typeof t.onFocus=="function"&&t.onFocus(d),typeof t.inputProps.onFocus=="function"&&t.inputProps.onFocus(d),g&&g.onFocus?g.onFocus():z(!0)}}))}}),y(()=>t.endAdornment),y(()=>Fe())]}}))]},"InputBase"));function He(e){return $("MuiFilledInput",e)}s(He,"getFilledInputUtilityClass");const L=R("MuiFilledInput",["root","colorSecondary","underline","focused","disabled","adornedStart","adornedEnd","error","sizeSmall","multiline","hiddenLabel","input","inputSizeSmall","inputHiddenLabel","inputMultiline","inputAdornedStart","inputAdornedEnd"]),Ee=F()({name:"MuiFilledInput",propDefaults:({set:e})=>e({components:{},fullWidth:!1,inputComponent:"input",multiline:!1,type:"text",hiddenLabel:!1}),selfPropNames:["classes","disableUnderline","hiddenLabel"],utilityClass:He,slotClasses:e=>({root:["root",!e.disableUnderline&&"underline"],input:["input"]})}),Ue=h(H,{name:"MuiFilledInput",slot:"Root",overridesResolver:(e,n)=>{const{ownerState:l}=e;return[...N(e,n),!l.disableUnderline&&n.underline]}})(({theme:e,ownerState:n})=>{const l=e.palette.mode==="light",t=l?"rgba(0, 0, 0, 0.42)":"rgba(255, 255, 255, 0.7)",r=l?"rgba(0, 0, 0, 0.06)":"rgba(255, 255, 255, 0.09)";return{position:"relative",backgroundColor:r,borderTopLeftRadius:`${e.shape.borderRadius}px`,borderTopRightRadius:`${e.shape.borderRadius}px`,transition:e.transitions.create("background-color",{duration:e.transitions.duration.shorter,easing:e.transitions.easing.easeOut}),"&:hover":{backgroundColor:l?"rgba(0, 0, 0, 0.09)":"rgba(255, 255, 255, 0.13)","@media (hover: none)":{backgroundColor:r}},[`&.${L.focused}`]:{backgroundColor:r},[`&.${L.disabled}`]:{backgroundColor:l?"rgba(0, 0, 0, 0.12)":"rgba(255, 255, 255, 0.12)"},...!n.disableUnderline&&{"&:after":{borderBottom:`2px solid ${e.palette[n.color].main}`,left:0,bottom:0,content:'""',position:"absolute",right:0,transform:"scaleX(0)",transition:e.transitions.create("transform",{duration:e.transitions.duration.shorter,easing:e.transitions.easing.easeOut}),pointerEvents:"none"},[`&.${L.focused}:after`]:{transform:"scaleX(1)"},[`&.${L.error}:after`]:{borderBottomColor:e.palette.error.main,transform:"scaleX(1)"},"&:before":{borderBottom:`1px solid ${t}`,left:0,bottom:0,content:'"\\00a0"',position:"absolute",right:0,transition:e.transitions.create("border-bottom-color",{duration:e.transitions.duration.shorter}),pointerEvents:"none"},[`&:hover:not(.${L.disabled}):before`]:{borderBottom:`1px solid ${e.palette.text.primary}`},[`&.${L.disabled}:before`]:{borderBottomStyle:"dotted"}},...n.startAdornment&&{paddingLeft:12},...n.endAdornment&&{paddingRight:12},...n.multiline&&{padding:"25px 12px 8px",...n.size==="small"&&{paddingTop:21,paddingBottom:4},...n.hiddenLabel&&{paddingTop:16,paddingBottom:17}}}}),De=h(E,{name:"MuiFilledInput",slot:"Input",overridesResolver:V})(({theme:e,ownerState:n})=>({paddingTop:25,paddingRight:12,paddingBottom:8,paddingLeft:12,"&:-webkit-autofill":{WebkitBoxShadow:e.palette.mode==="light"?null:"0 0 0 100px #266798 inset",WebkitTextFillColor:e.palette.mode==="light"?null:"#fff",caretColor:e.palette.mode==="light"?null:"#fff",borderTopLeftRadius:"inherit",borderTopRightRadius:"inherit"},...n.size==="small"&&{paddingTop:21,paddingBottom:4},...n.hiddenLabel&&{paddingTop:16,paddingBottom:17},...n.multiline&&{paddingTop:0,paddingBottom:0,paddingLeft:0,paddingRight:0},...n.startAdornment&&{paddingLeft:0},...n.endAdornment&&{paddingRight:0},...n.hiddenLabel&&n.size==="small"&&{paddingTop:8,paddingBottom:9}})),_e=Ee.component(s(function({allProps:n,classes:l,otherProps:t,props:r}){const a=b(()=>{const f={root:{ownerState:n},input:{ownerState:n}};return t.componentsProps?ce(t.componentsProps,f):f}),i=p(()=>r.classes||{},l);return o(Y,p(t,{get components(){return{Root:Ue,Input:De,...t.components||{}}},get componentsProps(){return a()},classes:i}))},"FilledInput"));function Ke(e){return $("MuiFormHelperText",e)}s(Ke,"getFormHelperTextUtilityClasses");const ie=R("MuiFormHelperText",["root","error","disabled","sizeSmall","sizeMedium","contained","focused","filled","required"]),Xe=q('<span class="notranslate">&#8203;</span>'),Ge=F()({name:"MuiFormHelperText",propDefaults:({set:e})=>e({component:"p"}),selfPropNames:["children","classes","disabled","error","filled","focused","margin","required","variant"],utilityClass:Ke,slotClasses:e=>({root:["root",!!e.disabled&&"disabled",!!e.error&&"error",!!e.size&&`size${P(e.size)}`,e.contained&&"contained",!!e.focused&&"focused",!!e.filled&&"filled",!!e.required&&"required"]})}),je=h("p",{name:"MuiFormHelperText",slot:"Root",overridesResolver:(e,n)=>{const{ownerState:l}=e;return[n.root,l.size&&n[`size${P(l.size)}`],l.contained&&n.contained,l.filled&&n.filled]}})(({theme:e,ownerState:n})=>({color:e.palette.text.secondary,...e.typography.caption,textAlign:"left",marginTop:3,marginRight:0,marginBottom:0,marginLeft:0,[`&.${ie.disabled}`]:{color:e.palette.text.disabled},[`&.${ie.error}`]:{color:e.palette.error.main},...n.size==="small"&&{marginTop:4},...n.contained&&{marginLeft:14,marginRight:14}})),Je=Ge.component(s(function({allProps:n,classes:l,otherProps:t,props:r}){const a=O(),i=w({props:n,muiFormControl:a,states:["variant","size","disabled","error","filled","focused","required"]}),f=p(n,()=>({contained:i.variant==="filled"||i.variant==="outlined",variant:i.variant,size:i.size,disabled:i.disabled,error:i.error,filled:i.filled,focused:i.focused,required:i.required})),c=Le(()=>r.children);return o(je,p(t,{get as(){return t.component},ownerState:f,get className(){return W(l.root,t.className)},get children(){return y(()=>c()===" ",!0)()?Xe.cloneNode(!0):c()}}))},"FormHelperText"));function Qe(e){return $("MuiInput",e)}s(Qe,"getInputUtilityClass");const M=R("MuiInput",["root","formControl","focused","disabled","colorSecondary","underline","error","sizeSmall","multiline","fullWidth","input","inputSizeSmall","inputMultiline","inputTypeSearch"]),Ye=F()({name:"MuiInput",propDefaults:({set:e})=>e({components:{},fullWidth:!1,inputComponent:"input",multiline:!1,type:"text"}),selfPropNames:["classes","disableUnderline"],utilityClass:Qe,slotClasses:e=>({root:["root",!e.disableUnderline&&"underline"],input:["input"]})}),Ze=h(H,{skipProps:J.filter(e=>e!=="classes"),name:"MuiInput",slot:"Root",overridesResolver:(e,n)=>{const{ownerState:l}=e;return[...N(e,n),!l.disableUnderline&&n.underline]}})(({theme:e,ownerState:n})=>{const t=e.palette.mode==="light"?"rgba(0, 0, 0, 0.42)":"rgba(255, 255, 255, 0.7)";return{position:"relative",...n.formControl&&{"label + &":{marginTop:16}},...!n.disableUnderline&&{"&:after":{borderBottom:`2px solid ${e.palette[n.color].main}`,left:0,bottom:0,content:'""',position:"absolute",right:0,transform:"scaleX(0)",transition:e.transitions.create("transform",{duration:e.transitions.duration.shorter,easing:e.transitions.easing.easeOut}),pointerEvents:"none"},[`&.${M.focused}:after`]:{transform:"scaleX(1)"},[`&.${M.error}:after`]:{borderBottomColor:e.palette.error.main,transform:"scaleX(1)"},"&:before":{borderBottom:`1px solid ${t}`,left:0,bottom:0,content:'"\\00a0"',position:"absolute",right:0,transition:e.transitions.create("border-bottom-color",{duration:e.transitions.duration.shorter}),pointerEvents:"none"},[`&:hover:not(.${M.disabled}):before`]:{borderBottom:`2px solid ${e.palette.text.primary}`,"@media (hover: none)":{borderBottom:`1px solid ${t}`}},[`&.${M.disabled}:before`]:{borderBottomStyle:"dotted"}}}}),en=h(E,{name:"MuiInput",slot:"Input",overridesResolver:V})({}),nn=Ye.component(s(function({classes:n,otherProps:l,props:t}){const r=b(()=>{const f={root:{ownerState:{disableUnderline:t.disableUnderline}}};return l.componentsProps?ce(l.componentsProps,f):f}),a=p(n,()=>t.classes||{});return o(Y,p(l,{get components(){return{Root:Ze,Input:en,...l.components||{}}},get componentsProps(){return r()},classes:a}))},"Input"));function tn(e){return $("MuiFormLabel",e)}s(tn,"getFormLabelUtilityClasses");const B=R("MuiFormLabel",["root","colorSecondary","focused","disabled","error","filled","required","asterisk"]),ae=F()({name:"MuiFormLabel",propDefaults:({set:e})=>e({component:"label"}),selfPropNames:["children","classes","color","disabled","error","filled","focused","required"],autoCallUseClasses:!1,utilityClass:tn,slotClasses:e=>({root:["root",`color${P(e.color)}`,!!e.disabled&&"disabled",!!e.error&&"error",!!e.filled&&"filled",!!e.focused&&"focused",!!e.required&&"required"],asterisk:["asterisk",!!e.error&&"error"]})}),ln=h("label",{name:"MuiFormLabel",slot:"Root",overridesResolver:({ownerState:e},n)=>({...n.root,...e.color==="secondary"&&n.colorSecondary,...e.filled&&n.filled})})(({theme:e,ownerState:n})=>({color:e.palette.text.secondary,...e.typography.body1,lineHeight:"1.4375em",padding:0,position:"relative",[`&.${B.focused}`]:{color:e.palette[n.color].main},[`&.${B.disabled}`]:{color:e.palette.text.disabled},[`&.${B.error}`]:{color:e.palette.error.main}})),rn=h("span",{name:"MuiFormLabel",slot:"Asterisk",overridesResolver:(e,n)=>n.asterisk})(({theme:e})=>({[`&.${B.error}`]:{color:e.palette.error.main}})),on=ae.component(s(function({allProps:n,otherProps:l,props:t}){const r=O(),a=w({props:n,muiFormControl:r,states:["color","required","focused","disabled","error","filled"]}),i=p(n,()=>({color:a.color||"primary",disabled:a.disabled,error:a.error,filled:a.filled,focused:a.focused,required:a.required})),f=ae.useClasses(i);return o(ln,p(l,{get as(){return l.component},ownerState:i,get className(){return W(f.root,l.className)},get children(){return[y(()=>t.children),o(pe,{get when(){return a.required},get children(){return o(rn,{ownerState:i,"aria-hidden":!0,get className(){return f.asterisk},get children(){return["\u2009","*"]}})}})]}}))},"FormLabel"));function an(e){return $("MuiInputLabel",e)}s(an,"getInputLabelUtilityClasses");R("MuiInputLabel",["root","focused","disabled","error","required","asterisk","formControl","sizeSmall","shrink","animated","standard","filled","outlined"]);const de=F()({name:"MuiInputLabel",propDefaults:({set:e})=>e({disableAnimation:!1}),selfPropNames:["children","classes","color","disableAnimation","disabled","error","focused","margin","required","shrink","variant"],autoCallUseClasses:!1,utilityClass:an,slotClasses:e=>({root:["root",!!e.formControl&&"formControl",!e.disableAnimation&&"animated",!!e.shrink&&"shrink",e.size==="small"&&"sizeSmall",!!e.variant&&e.variant],asterisk:[!!e.required&&"asterisk"]})}),dn=h(on,{skipProps:J.filter(e=>e!=="classes"),name:"MuiInputLabel",slot:"Root",overridesResolver:(e,n)=>{const{ownerState:l}=e;return[{[`& .${B.asterisk}`]:n.asterisk},n.root,l.formControl&&n.formControl,l.size==="small"&&n.sizeSmall,l.shrink&&n.shrink,!l.disableAnimation&&n.animated,n[l.variant]]}})(({theme:e,ownerState:n})=>({display:"block",transformOrigin:"top left",whiteSpace:"nowrap",overflow:"hidden",textOverflow:"ellipsis",maxWidth:"100%",...n.formControl&&{position:"absolute",left:0,top:0,transform:"translate(0, 20px) scale(1)"},...n.size==="small"&&{transform:"translate(0, 17px) scale(1)"},...n.shrink&&{transform:"translate(0, -1.5px) scale(0.75)",transformOrigin:"top left",maxWidth:"133%"},...!n.disableAnimation&&{transition:e.transitions.create(["color","transform","max-width"],{duration:e.transitions.duration.shorter,easing:e.transitions.easing.easeOut})},...n.variant==="filled"&&{zIndex:1,pointerEvents:"none",transform:"translate(12px, 16px) scale(1)",maxWidth:"calc(100% - 24px)",...n.size==="small"&&{transform:"translate(12px, 13px) scale(1)"},...n.shrink&&{userSelect:"none",pointerEvents:"auto",transform:"translate(12px, 7px) scale(0.75)",maxWidth:"calc(133% - 24px)",...n.size==="small"&&{transform:"translate(12px, 4px) scale(0.75)"}}},...n.variant==="outlined"&&{zIndex:1,pointerEvents:"none",transform:"translate(14px, 16px) scale(1)",maxWidth:"calc(100% - 24px)",...n.size==="small"&&{transform:"translate(14px, 9px) scale(1)"},...n.shrink&&{userSelect:"none",pointerEvents:"auto",maxWidth:"calc(133% - 24px)",transform:"translate(14px, -9px) scale(0.75)"}}})),sn=de.component(s(function({allProps:n,props:l}){const t=O(),[,r]=ze(n,["disableAnimation","margin","shrink","variant"]),a=b(()=>{let I=l.shrink;return typeof I>"u"&&t&&(I=t.filled||t.focused||t.adornedStart),I}),i=w({props:n,muiFormControl:t,states:["size","variant","required"]}),f=p(n,()=>({formControl:t,shrink:a(),size:i.size,variant:i.variant,required:i.required})),c=de.useClasses(f),k=p(c,()=>l.classes||{});return o(dn,p(r,{get["data-shrink"](){return a()},ownerState:f,classes:k}))},"InputLabel")),un=q("<span></span>"),cn=q('<span class="notranslate">&#8203;</span>'),pn=F()({name:"MuiNotchedOutline",selfPropNames:["disabled","error","focused","label","notched"]}),mn=h("fieldset")({textAlign:"left",position:"absolute",bottom:0,right:0,top:-5,left:0,margin:0,padding:"0 8px",pointerEvents:"none",borderRadius:"inherit",borderStyle:"solid",borderWidth:"1px",overflow:"hidden",minWidth:"0%"}),fn=h("legend")(({ownerState:e,theme:n})=>({float:"unset",...!e.withLabel&&{padding:0,lineHeight:"11px",transition:n.transitions.create("width",{duration:150,easing:n.transitions.easing.easeOut})},...e.withLabel&&{display:"block",width:"auto",padding:0,height:11,fontSize:"0.75em",visibility:"hidden",maxWidth:"0.01px",transition:n.transitions.create("max-width",{duration:50,easing:n.transitions.easing.easeOut}),whiteSpace:"nowrap","& > span":{paddingLeft:5,paddingRight:5,display:"inline-block"},...e.notched&&{maxWidth:"100%",transition:n.transitions.create("max-width",{duration:100,easing:n.transitions.easing.easeOut,delay:50})}}})),bn=pn.component(s(function({allProps:n,otherProps:l,props:t}){const r=b(()=>t.label!=null&&t.label!==""),a=p(n,()=>({withLabel:r()}));return o(mn,p(l,{"aria-hidden":!0,get className(){return l.className},ownerState:a,get children(){return o(fn,{ownerState:a,get children(){return o(pe,{get when(){return r()},get fallback(){return cn.cloneNode(!0)},get children(){const i=un.cloneNode(!0);return m(i,()=>t.label),i}})}})}}))},"NotchedOutline"));function gn(e){return $("MuiOutlinedInput",e)}s(gn,"getOutlinedInputUtilityClass");const T=R("MuiOutlinedInput",["root","colorSecondary","focused","disabled","adornedStart","adornedEnd","error","sizeSmall","multiline","notchedOutline","input","inputSizeSmall","inputMultiline","inputAdornedStart","inputAdornedEnd"]),hn=F()({name:"MuiOutlinedInput",propDefaults:({set:e})=>e({components:{},fullWidth:!1,inputComponent:"input",multiline:!1,type:"text"}),selfPropNames:["classes","label","notched"],utilityClass:gn,slotClasses:()=>({root:["root"],notchedOutline:["notchedOutline"],input:["input"]})}),xn=h(H,{skipProps:J.filter(e=>e!=="classes"),name:"MuiOutlinedInput",slot:"Root",overridesResolver:N})(({theme:e,ownerState:n})=>{const l=e.palette.mode==="light"?"rgba(0, 0, 0, 0.23)":"rgba(255, 255, 255, 0.23)";return{position:"relative",borderRadius:e.shape.borderRadius,[`&:hover .${T.notchedOutline}`]:{borderColor:e.palette.text.primary},"@media (hover: none)":{[`&:hover .${T.notchedOutline}`]:{borderColor:l}},[`&.${T.focused} .${T.notchedOutline}`]:{borderColor:e.palette[n.color].main,borderWidth:"2px"},[`&.${T.error} .${T.notchedOutline}`]:{borderColor:e.palette.error.main},[`&.${T.disabled} .${T.notchedOutline}`]:{borderColor:e.palette.action.disabled},...n.startAdornment&&{paddingLeft:14},...n.endAdornment&&{paddingRight:14},...n.multiline&&{padding:"16.5px 14px",...n.size==="small"&&{padding:"8.5px 14px"}}}}),vn=h(bn,{name:"MuiOutlinedInput",slot:"NotchedOutline",overridesResolver:(e,n)=>n.notchedOutline})(({theme:e})=>({borderColor:e.palette.mode==="light"?"rgba(0, 0, 0, 0.23)":"rgba(255, 255, 255, 0.23)"})),Cn=h(E,{name:"MuiOutlinedInput",slot:"Input",overridesResolver:V})(({theme:e,ownerState:n})=>({padding:"16.5px 14px","&:-webkit-autofill":{WebkitBoxShadow:e.palette.mode==="light"?null:"0 0 0 100px #266798 inset",WebkitTextFillColor:e.palette.mode==="light"?null:"#fff",caretColor:e.palette.mode==="light"?null:"#fff",borderRadius:"inherit"},...n.size==="small"&&{padding:"8.5px 14px"},...n.multiline&&{padding:0},...n.startAdornment&&{paddingLeft:0},...n.endAdornment&&{paddingRight:0}})),yn=hn.component(s(function({allProps:n,classes:l,otherProps:t,props:r}){const a=O(),i=w({props:n,muiFormControl:a,states:["required"]}),f=b(()=>{const c=r.label;return c!=null&&c!==""&&i.required?[c,"\xA0","*"]:c});return o(Y,p({renderSuffix:c=>o(vn,{get className(){return l.notchedOutline},get label(){return f()},get notched(){return y(()=>typeof r.notched<"u",!0)()?r.notched:Boolean(c.startAdornment||c.filled||c.focused)}})},t,{get components(){return{Root:xn,Input:Cn,...t.components||{}}},get classes(){return{...l,...r.classes,notchedOutline:null}}}))},"OutlinedInput"));function Fn(e){return $("MuiTextField",e)}s(Fn,"getTextFieldUtilityClass");R("MuiTextField",["root"]);function In(e){return b(()=>e?.()??`mui-${Be()}`)}s(In,"createUniqueId");const Tn=F()({name:"MuiTextField",propDefaults:({set:e})=>e({autoFocus:!1,color:"primary",disabled:!1,error:!1,fullWidth:!1,multiline:!1,required:!1,select:!1,variant:"outlined"}),selfPropNames:["FormHelperTextProps","InputLabelProps","InputProps","autoComplete","autoFocus","children","classes","color","defaultValue","disabled","error","fullWidth","helperText","id","inputProps","inputRef","label","maxRows","minRows","multiline","name","onBlur","onChange","onFocus","placeholder","required","rows","select","size","type","value","variant"],utilityClass:Fn,slotClasses:()=>({root:["root"]})}),Rn={standard:nn,filled:_e,outlined:yn},$n=h(Pe,{name:"MuiTextField",slot:"Root",overridesResolver:(e,n)=>n.root})({}),u=Tn.component(s(function({allProps:n,classes:l,otherProps:t,props:r}){const a=b(()=>{const x={};return r.variant==="outlined"&&(r.InputLabelProps&&typeof r.InputLabelProps.shrink<"u"&&(x.notched=r.InputLabelProps.shrink),x.label=r.label),x}),i=In(()=>r.id),f=b(()=>r.helperText&&i()?`${i()}-helper-text`:void 0),c=b(()=>r.label&&i()?`${i()}-label`:void 0),k=b(()=>Rn[r.variant]),I=s(()=>o(G,p({get component(){return k()},get["aria-describedby"](){return f()},get autoComplete(){return r.autoComplete},get autoFocus(){return r.autoFocus},get defaultValue(){return r.defaultValue},get fullWidth(){return r.fullWidth},get multiline(){return r.multiline},get name(){return r.name},get rows(){return r.rows},get maxRows(){return r.maxRows},get minRows(){return r.minRows},get type(){return r.type},get value(){return r.value},get id(){return i()},get inputRef(){return r.inputRef},get onBlur(){return r.onBlur},get onChange(){return r.onChange},get onFocus(){return r.onFocus},get placeholder(){return r.placeholder},get inputProps(){return r.inputProps}},a,()=>r.InputProps||{})),"InputElement"),z=s(()=>{const x=r.label;return x!=null&&x!==""&&o(sn,p({get htmlFor(){return i()},get id(){return c()}},()=>r.InputLabelProps||{},{children:x}))},"label"),g=s(()=>{const x=r.helperText;if(x)return o(Je,p({get id(){return f()}},()=>r.FormHelperTextProps||{},{children:x}))},"helperText");return o($n,p({get className(){return W(l.root,t.className)},get disabled(){return r.disabled},get error(){return r.error},get fullWidth(){return r.fullWidth},get required(){return r.required},get color(){return r.color},get variant(){return r.variant},ownerState:n},t,{get children(){return[y(()=>z()),y(()=>I()),y(()=>g())]}}))},"TextField"));function me(){return o(Q,{component:"form",sx:{"& > :not(style)":{m:1,width:"25ch"}},noValidate:!0,autocomplete:"off",get children(){return[o(u,{id:"outlined-basic",label:"Outlined",variant:"outlined",fullWidth:!0}),o(u,{id:"filled-basic",label:"Filled",variant:"filled"}),o(u,{id:"standard-basic",label:"Standard",variant:"standard"})]}})}s(me,"BasicTextFields");me.code=`import Box from "@suid/material/Box";
+import TextField from "@suid/material/TextField";
+
+export default function BasicTextFields() {
+  return (
+    <Box
+      component="form"
+      sx={{
+        "& > :not(style)": { m: 1, width: "25ch" },
+      }}
+      noValidate
+      autocomplete="off"
+    >
+      <TextField
+        id="outlined-basic"
+        label="Outlined"
+        variant="outlined"
+        fullWidth
+      />
+      <TextField id="filled-basic" label="Filled" variant="filled" />
+      <TextField id="standard-basic" label="Standard" variant="standard" />
+    </Box>
+  );
+}
+`;const X=q("<div></div>");function fe(){return o(Q,{component:"form",sx:{maxWidth:764,[`& .${u}`]:{m:1,width:"25ch"}},noValidate:!0,autocomplete:"off",get children(){return[(()=>{const e=X.cloneNode(!0);return m(e,o(u,{required:!0,id:"outlined-required",label:"Required",defaultValue:"Hello World"}),null),m(e,o(u,{disabled:!0,id:"outlined-disabled",label:"Disabled",defaultValue:"Hello World"}),null),m(e,o(u,{id:"outlined-password-input",label:"Password",type:"password",autoComplete:"current-password"}),null),m(e,o(u,{id:"outlined-read-only-input",label:"Read Only",defaultValue:"Hello World",InputProps:{readOnly:!0}}),null),m(e,o(u,{id:"outlined-number",label:"Number",type:"number",InputLabelProps:{shrink:!0}}),null),m(e,o(u,{id:"outlined-search",label:"Search field",type:"search"}),null),m(e,o(u,{id:"outlined-helperText",label:"Helper text",defaultValue:"Default Value",helperText:"Some important text"}),null),e})(),(()=>{const e=X.cloneNode(!0);return m(e,o(u,{required:!0,id:"filled-required",label:"Required",defaultValue:"Hello World",variant:"filled"}),null),m(e,o(u,{disabled:!0,id:"filled-disabled",label:"Disabled",defaultValue:"Hello World",variant:"filled"}),null),m(e,o(u,{id:"filled-password-input",label:"Password",type:"password",autoComplete:"current-password",variant:"filled"}),null),m(e,o(u,{id:"filled-read-only-input",label:"Read Only",defaultValue:"Hello World",InputProps:{readOnly:!0},variant:"filled"}),null),m(e,o(u,{id:"filled-number",label:"Number",type:"number",InputLabelProps:{shrink:!0},variant:"filled"}),null),m(e,o(u,{id:"filled-search",label:"Search field",type:"search",variant:"filled"}),null),m(e,o(u,{id:"filled-helperText",label:"Helper text",defaultValue:"Default Value",helperText:"Some important text",variant:"filled"}),null),e})(),(()=>{const e=X.cloneNode(!0);return m(e,o(u,{required:!0,id:"standard-required",label:"Required",defaultValue:"Hello World",variant:"standard"}),null),m(e,o(u,{disabled:!0,id:"standard-disabled",label:"Disabled",defaultValue:"Hello World",variant:"standard"}),null),m(e,o(u,{id:"standard-password-input",label:"Password",type:"password",autoComplete:"current-password",variant:"standard"}),null),m(e,o(u,{id:"standard-read-only-input",label:"Read Only",defaultValue:"Hello World",InputProps:{readOnly:!0},variant:"standard"}),null),m(e,o(u,{id:"standard-number",label:"Number",type:"number",InputLabelProps:{shrink:!0},variant:"standard"}),null),m(e,o(u,{id:"standard-search",label:"Search field",type:"search",variant:"standard"}),null),m(e,o(u,{id:"standard-helperText",label:"Helper text",defaultValue:"Default Value",helperText:"Some important text",variant:"standard"}),null),e})()]}})}s(fe,"FormPropsTextFields");fe.code=`import Box from "@suid/material/Box";
+import TextField from "@suid/material/TextField";
+
+export default function FormPropsTextFields() {
+  return (
+    <Box
+      component="form"
+      sx={{
+        maxWidth: 764,
+        [\`& .\${TextField}\`]: { m: 1, width: "25ch" },
+      }}
+      noValidate
+      autocomplete="off"
+    >
+      <div>
+        <TextField
+          required
+          id="outlined-required"
+          label="Required"
+          defaultValue="Hello World"
+        />
+        <TextField
+          disabled
+          id="outlined-disabled"
+          label="Disabled"
+          defaultValue="Hello World"
+        />
+        <TextField
+          id="outlined-password-input"
+          label="Password"
+          type="password"
+          autoComplete="current-password"
+        />
+        <TextField
+          id="outlined-read-only-input"
+          label="Read Only"
+          defaultValue="Hello World"
+          InputProps={{
+            readOnly: true,
+          }}
+        />
+        <TextField
+          id="outlined-number"
+          label="Number"
+          type="number"
+          InputLabelProps={{
+            shrink: true,
+          }}
+        />
+        <TextField id="outlined-search" label="Search field" type="search" />
+        <TextField
+          id="outlined-helperText"
+          label="Helper text"
+          defaultValue="Default Value"
+          helperText="Some important text"
+        />
+      </div>
+      <div>
+        <TextField
+          required
+          id="filled-required"
+          label="Required"
+          defaultValue="Hello World"
+          variant="filled"
+        />
+        <TextField
+          disabled
+          id="filled-disabled"
+          label="Disabled"
+          defaultValue="Hello World"
+          variant="filled"
+        />
+        <TextField
+          id="filled-password-input"
+          label="Password"
+          type="password"
+          autoComplete="current-password"
+          variant="filled"
+        />
+        <TextField
+          id="filled-read-only-input"
+          label="Read Only"
+          defaultValue="Hello World"
+          InputProps={{
+            readOnly: true,
+          }}
+          variant="filled"
+        />
+        <TextField
+          id="filled-number"
+          label="Number"
+          type="number"
+          InputLabelProps={{
+            shrink: true,
+          }}
+          variant="filled"
+        />
+        <TextField
+          id="filled-search"
+          label="Search field"
+          type="search"
+          variant="filled"
+        />
+        <TextField
+          id="filled-helperText"
+          label="Helper text"
+          defaultValue="Default Value"
+          helperText="Some important text"
+          variant="filled"
+        />
+      </div>
+      <div>
+        <TextField
+          required
+          id="standard-required"
+          label="Required"
+          defaultValue="Hello World"
+          variant="standard"
+        />
+        <TextField
+          disabled
+          id="standard-disabled"
+          label="Disabled"
+          defaultValue="Hello World"
+          variant="standard"
+        />
+        <TextField
+          id="standard-password-input"
+          label="Password"
+          type="password"
+          autoComplete="current-password"
+          variant="standard"
+        />
+        <TextField
+          id="standard-read-only-input"
+          label="Read Only"
+          defaultValue="Hello World"
+          InputProps={{
+            readOnly: true,
+          }}
+          variant="standard"
+        />
+        <TextField
+          id="standard-number"
+          label="Number"
+          type="number"
+          InputLabelProps={{
+            shrink: true,
+          }}
+          variant="standard"
+        />
+        <TextField
+          id="standard-search"
+          label="Search field"
+          type="search"
+          variant="standard"
+        />
+        <TextField
+          id="standard-helperText"
+          label="Helper text"
+          defaultValue="Default Value"
+          helperText="Some important text"
+          variant="standard"
+        />
+      </div>
+    </Box>
+  );
+}
+`;function se(e){return e.split(" ").map(n=>n.slice(0,1).toUpperCase()+n.slice(1)).join(" ")}s(se,"capitalizeWords");function be(){const[e,n]=ue(se("Cat in the Hat"));return o(Q,{component:"form",sx:{"& > :not(style)":{m:1,width:"25ch"}},novalidate:!0,autocomplete:"off",get children(){return[o(u,{id:"outlined-name",label:"Name",get value(){return e()},onChange:(l,t)=>{n(se(t))}}),o(u,{id:"outlined-uncontrolled",label:"Uncontrolled",defaultValue:"foo"})]}})}s(be,"StateTextFields");be.code=`import Box from "@suid/material/Box";
+import TextField from "@suid/material/TextField";
+import { createSignal } from "solid-js";
+
+function capitalizeWords(value: string) {
+  return value
+    .split(" ")
+    .map((v) => v.slice(0, 1).toUpperCase() + v.slice(1))
+    .join(" ");
+}
+
+export default function StateTextFields() {
+  const [name, setName] = createSignal(capitalizeWords("Cat in the Hat"));
+
+  return (
+    <Box
+      component="form"
+      sx={{
+        "& > :not(style)": { m: 1, width: "25ch" },
+      }}
+      novalidate
+      autocomplete="off"
+    >
+      <TextField
+        id="outlined-name"
+        label="Name"
+        value={name()}
+        onChange={(event, value) => {
+          setName(capitalizeWords(value));
+        }}
+      />
+      <TextField
+        id="outlined-uncontrolled"
+        label="Uncontrolled"
+        defaultValue="foo"
+      />
+    </Box>
+  );
+}
+`;function Sn(){return o(we,{get name(){return u.name},docsName:"text-fields",examples:[me,fe,be]})}s(Sn,"TextFieldPage");export{Sn as default};